requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
gunicorn==19.0.0
itsdangerous==0.24
wsgiref==0.1.2
boto==2.15.0
requests==2.2.1
pymongo==2.6.3
webassets==0.10.1
Flask-Assets==0.10
yuicompressor==2.4.8
git+https://github.com/dchaplinsky/unshred.git
Flask-Login==0.2.11
python-social-auth==0.1.26
flask-mongoengine==0.7.0
click==2.4
